cleaning up lint and making sure tests are passing
