+++ pdr_backend/lake/persistent_data_store.py
@@ -64,9 +64,9 @@ class PersistentDataStore(BaseDataStore):
         """
 
         views = self.duckdb_conn.execute("SELECT * FROM duckdb_views;").fetchall()
-        view_names = [view [4] for view in views]
+        view_names = [view[4] for view in views]
         return view_names
-    
+
     @enforce_types
     def table_exists(self, table_name: str):
         table_names = self.get_table_names()
