+++ pdr_backend/lake/test/test_table_bronze_pdr_slots.py
@@ -63,18 +63,10 @@ def test_table_bronze_pdr_slots(
     }
 
     # Work 1: Append all data onto tables
-    gql_tables["pdr_predictions"].append_to_storage(
-        _gql_datafactory_etl_predictions_df
-    )
-    gql_tables["pdr_truevals"].append_to_storage(
-        _gql_datafactory_etl_truevals_df
-    )
-    gql_tables["pdr_payouts"].append_to_storage(
-        _gql_datafactory_etl_payouts_df
-    )
-    gql_tables["pdr_slots"].append_to_storage(
-        _gql_datafactory_etl_slots_df
-    )
+    gql_tables["pdr_predictions"].append_to_storage(_gql_datafactory_etl_predictions_df)
+    gql_tables["pdr_truevals"].append_to_storage(_gql_datafactory_etl_truevals_df)
+    gql_tables["pdr_payouts"].append_to_storage(_gql_datafactory_etl_payouts_df)
+    gql_tables["pdr_slots"].append_to_storage(_gql_datafactory_etl_slots_df)
 
     # Check that the data is appended correctly
     pds = PersistentDataStore(ppss.lake_ss.lake_dir)
